+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.35.1"
-__released__ = "2023-04-08"
+__version__ = "1.35.2"
+__released__ = "2023-04-21"
 
 from .emborg import Emborg